requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0

langchain>=0.2.14
langchain-openai>=0.1.22
langchain-community>=0.2.12
langgraph>=0.2.40

chromadb>=0.5.5
langchain-chroma>=0.1.0

pdfplumber>=0.11.4
pypdf>=4.3.1
tiktoken>=0.7.0
